requirements.txt
python-dotenv==1.0.0
flask==3.0.2
flask-cors==4.0.0
requests==2.31.0
google-generativeai==0.3.2
gunicorn==21.2.0 